+++ app/src/main/java/com/ugb/miprimercalculadora/MainActivity.java
@@ -20,6 +20,7 @@ import android.view.MenuItem;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.Button;
+import android.widget.EditText;
 import android.widget.ListView;
 import android.widget.RadioButton;
 import android.widget.RelativeLayout;
@@ -31,11 +32,19 @@ import android.widget.Toast;
 import com.firebase.ui.auth.AuthUI;
 import com.google.android.gms.auth.GoogleAuthUtil;
 import com.google.android.gms.auth.api.Auth;
+import com.google.android.gms.auth.api.signin.GoogleSignIn;
+import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
+import com.google.android.gms.auth.api.signin.GoogleSignInClient;
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
+import com.google.android.gms.common.api.ApiException;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import com.google.firebase.auth.AuthCredential;
+import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.auth.GoogleAuthProvider;
 import com.google.firebase.auth.OAuthProvider;
 
 import org.json.JSONArray;
@@ -55,6 +64,11 @@ import java.util.List;
 import java.util.MissingFormatArgumentException;
 
 public class MainActivity extends AppCompatActivity {
+    private FirebaseAuth mAuth;
+    public static final int RC_SING_In = 9876;
+    private FirebaseAuth.AuthStateListener mAuthListener;
+    private GoogleSignInClient mGoogleSignClien;
+
     Button btn;
 
 
@@ -62,40 +76,92 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+        //Firebase Inicializacion
+        mAuth = FirebaseAuth.getInstance();
 
+        //Pantallita Google
+        createResquet();
 
-
-
-
-
-
-
-
-        //Iniciar sesion
-        btn = findViewById(R.id.btnIniciar);
+        btn = findViewById(R.id.btnGoogle);
         btn.setOnClickListener(v -> {
-            IniciarSesion();
+            signIn();
         });
 
-        //Registrar
-        btn = findViewById(R.id.btnRegistrar);
-        btn.setOnClickListener(v -> {
-            Resgistrarse();
-        });
 
 
     }
 
-    //iniciar Sesion
-    private void IniciarSesion(){
-        Intent iniciarSesion = new Intent(this, IniciarSesion.class);
-        startActivity(iniciarSesion);
+    //Creacion de Token
+    private void createResquet(){
+        GoogleSignInOptions gso = new GoogleSignInOptions
+                .Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
+                .requestIdToken(getString(R.string.default_web_client_id))
+                .requestEmail()
+                .build();
+        mGoogleSignClien = GoogleSignIn.getClient(this, gso);
+
     }
 
-    //Registrar
-    private void Resgistrarse (){
-        Intent Registrar = new Intent(this, Registrar.class);
-        startActivity(Registrar);
+    //Metodo SignIn
+
+    private void signIn(){
+        Intent signIntent = mGoogleSignClien.getSignInIntent();
+        startActivityForResult(signIntent, RC_SING_In);
+    }
+
+    //Activity Result
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+
+        if (requestCode == RC_SING_In){
+            Task<GoogleSignInAccount> task = GoogleSignIn.getSignedInAccountFromIntent(data);
+            try {
+                GoogleSignInAccount account = task.getResult(ApiException.class);
+                firebaseAuthWithGoogle(account);
+
+            }catch (Exception e){
+
+            }
+        }
+    }
+
+  //Metodo de FirabaseAuth para google
+    private void firebaseAuthWithGoogle(GoogleSignInAccount acct) {
+        AuthCredential credential = GoogleAuthProvider.getCredential(acct.getIdToken(), null);
+        mAuth.signInWithCredential(credential)
+                .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
+                    @Override
+                    public void onComplete(@NonNull Task<AuthResult> task) {
+                        if (task.isSuccessful()) {
+                            // Sign in success, update UI with the signed-in user's information
+
+                            FirebaseUser user = mAuth.getCurrentUser();
+                            Intent home = new Intent(getApplicationContext(), Home.class);
+                            startActivity(home);
+                            Toast.makeText(getApplicationContext(), "Bienvenido a Mandaditos", Toast.LENGTH_LONG).show();
+
+                        } else {
+                            // If sign in fails, display a message to the user.
+                            Toast.makeText(getApplicationContext(), "Error de inicio", Toast.LENGTH_LONG).show();
+                        }
+
+                        // ...
+                    }
+                });
+    }
+
+    //OnStar para mantener sesion iniciada
+    @Override
+    public void onStart() {
+        super.onStart();
+        // Check if user is signed in (non-null) and update UI accordingly.
+        FirebaseUser currentUser = mAuth.getCurrentUser();
+        //updateUI(currentUser);
+        if (currentUser != null){
+            Intent home = new Intent(getApplicationContext(), Home.class);
+            startActivity(home);
+        }
     }
